backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
pydantic-settings==2.5.2
python-multipart==0.0.12
llama-index-core==0.12.0
llama-index-readers-file==0.4.0
dashscope==1.20.0
openai==1.50.0
openpyxl==3.1.5
python-dotenv==1.0.1
python-docx==1.1.2
aiofiles==24.1.0
httpx==0.27.0
pytest==8.3.0
pytest-asyncio==0.24.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
